Move upload method into files api module
